refactor(lineage): move core models into domain
