+++ yak-ops-business/yak-ops-business-lineage/src/main/java/io/yak/ops/business/lineage/domain/LineageRelationType.java
@@ -1,4 +1,4 @@
-package io.yak.ops.business.lineage;
+package io.yak.ops.business.lineage.domain;
 
 /**
  * Typed metadata relationships. Every edge is persisted from upstream asset to downstream asset.
